python/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
spacy
google-generativeai
transformers
tensorflow
selenium
webdriver-manager
pymupdf